[Webapp] display/create cards and variants
+ hierarchy: to see the whole project at a glance
+ project: friendly overview
+ initial (and minimal) card edition
+ ability to create custom cardDef

+++ colab-api/src/main/java/ch/colabproject/colab/api/persistence/card/CardDao.java
@@ -8,6 +8,7 @@ package ch.colabproject.colab.api.persistence.card;
 
 import ch.colabproject.colab.api.exceptions.ColabMergeException;
 import ch.colabproject.colab.api.model.card.Card;
+import ch.colabproject.colab.generator.model.exceptions.HttpErrorMessage;
 import java.util.List;
 import javax.ejb.LocalBean;
 import javax.ejb.Stateless;
@@ -114,6 +115,10 @@ public class CardDao {
         logger.debug("delete card #{}", id);
         // TODO: move to recycle bin first
         Card card = this.getCard(id);
+        if (card.getRootCardProject() != null){
+            // no way to delete the root card
+            throw HttpErrorMessage.badRequest();
+        }
         em.remove(card);
         return card;
     }
